% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.r
\name{apply_fun}
\alias{apply_fun}
\title{apply_fun}
\usage{
apply_fun(f, x)
}
\arguments{
\item{f}{a function}

\item{x}{an argument}
}
\description{
from https://ro-che.info/articles/2017-08-18-call-r-function-from-c
}
