% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.r
\name{eapply3}
\alias{eapply3}
\title{eapply3}
\usage{
eapply3(X, FUN, ...)
}
\description{
eapply3(X = xx,FUN = function(x=2,y=3,...){print(x);print(y);nest(...)},z=1)
}
\details{
X: environment, FUN function, ...
}
