% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hash_bucket}
\alias{hash_bucket}
\title{Hash Bucket From Hash Table}
\usage{
hash_bucket(obj, ht)
}
\arguments{
\item{obj}{index of hash bucket as an object that can be coerced to integer, or a string}

\item{ht}{'hash_table' object}
}
\value{
character vector
}
\description{
Returns the full bucket contents identified  by index or a string from the hash table.
}

