% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hash_index}
\alias{hash_index}
\title{Hash Index for String from Hash Table}
\usage{
hash_index(str, ht)
}
\arguments{
\item{str}{string under consideration}

\item{ht}{'hash_table' object}
}
\value{
integer
}
\description{
Returns the index of the bucket for the specified string and 'hash_table' object.
}

