% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_hashed}
\alias{is_hashed}
\title{Anser the question,"Is this environment hashed?"}
\usage{
is_hashed(env)
}
\arguments{
\item{env}{An environment.}
}
\value{
TRUE or FALSE
}
\description{
Anser the question,"Is this environment hashed?"
}

