% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.hash_table}
\alias{print.hash_table}
\title{Print Summary of Class hash_table}
\usage{
\method{print}{hash_table}(ht, ...)
}
\arguments{
\item{object}{of class 'hash_table' as returned by 'hash_table()'.}
}
\value{
the argument ht invisibly
}
\description{
Print a summary of the hash_table object
}

