% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_hashed.r
\name{is_hashed}
\alias{is_hashed}
\title{Anser the question,"Is this environment hashed?"}
\usage{
is_hashed(env)
}
\arguments{
\item{env}{An environment.}
}
\value{
TRUE or FALSE
}
\description{
Anser the question,"Is this environment hashed?"
}
